geoip: store GeoIP files in $(PREFIX)/share by default
The /etc directory shouldn't contain non-human-readable files.
netsniff-ng (when called with the '-U' option) currently installs the
GeoIP database files to /etc/netsniff-ng by default.

Change this to install them to $(PREFIX)/share/netsniff-ng instead,
which is conformant to the FHS [1].

[1] https://wiki.debian.org/FilesystemHierarchyStandard

Also create the respective directory in the 'make install' target.

Fixes #187

Signed-off-by: Tobias Klauser <tklauser@distanz.ch>