trafgen: Check pointer in ipv6 device address lookup function
The getifaddrs function returns a linked list node for L2
even in the case of L3 interfaces (e.g. a TUN device), but
with the pointer to the address structure set to NULL.
This causes a segfault when trying to access the address
family from it when such network interfaces are present
in the system.

The issue is solved by checking the validity of the pointer
beforehand: if the address structure is present the address
family can be fetched and the current iteration can go on
safely, otherwise the node is skipped (as it is not linked
to any address).

Signed-off-by: Matteo Perin <matteoperin1302@gmail.com>