Running trafgen on allowed CPUs only
Trafgen fails with the message "Can't set this CPU affinity!"
when the allowed CPUs are limited and the allocated cores
do not form a zero-based continuous range.
This patch adds relative-to-physical CPU mapping, allowing
trafgen to run in such environments. The command-line parameters
and packet configuration files use relative CPU numbers,
which are then mapped to the physical cores allocated to the task.

Signed-off-by: Pavel Dubovitsky <pdubovitsky@yahoo.com>